pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo9"
version = "1.0.9"
description = "Package version of WongKinYiu/yolov9"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython",
    "ipython",
    "matplotlib>=3.2.2",
    "numpy>=1.18.5",
    "opencv-python>=4.1.1",
    "Pillow>=7.1.2",
    "psutil",
    "PyYAML>=5.3.1",
    "requests>=2.23.0",
    "scipy>=1.4.1",
    "thop>=0.1.1",
    "torch>=1.7.0",
    "torchvision>=0.8.1",
    "tqdm>=4.64.0",
    "pandas>=1.1.4",
    "seaborn>=0.11.0",
    "albumentations>=1.0.3",
    "pycocotools>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.setuptools.package-data]
"*" = ["*"]
